Séparation en 2 classes, puisque la méthode calculSequenceFibonacci est à l'intérieur du switch case. J'étais pas sure de comment le séparé et reste dans le main
Premier Scanner supprimer puisqu'il était inutilisé.
